requirements.txt
translate-toolkit
lxml
pandas
openpyxl
